#74-가입 거절시 파일 삭제 구현
