+++ src/main/java/com/fofo/core/domain/image/FileStore.java
@@ -1,22 +1,36 @@
 package com.fofo.core.domain.image;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
+@Slf4j
 @Component
 @RequiredArgsConstructor
 public class FileStore {
 
     @Value("${file.dir}")
     private String fileDir;
+
+    public void deleteFile(final String filename) {
+        File existsFile = new File(getFullPath(filename));
+        Path exsitsPath = existsFile.toPath();
+        try {
+            Files.delete(exsitsPath);
+        } catch (IOException e) {
+            log.warn("파일 삭제에 실패했습니다.", e);
+        }
+    }
 
     public String getFullPath(final String filename) {
         String memberId = extractMemberId(filename);
@@ -65,4 +79,5 @@ public class FileStore {
         int pos = originalFilename.lastIndexOf(".");
         return originalFilename.substring(pos + 1);
     }
+
 }
